+++ src/api/consumer/lib/get-driver.js
@@ -1,0 +1,7 @@
+/** @flow */
+import { loadConsumer } from '../../../consumer';
+
+export default function getDriver() {
+  return loadConsumer()
+    .then(consumer => consumer.driver);
+}