+++ src/cli/command-registrar-builder.js
@@ -28,6 +28,7 @@ import CiUpdate from './commands/private-cmds/ci-update-cmd';
 import RefreshScope from './commands/private-cmds/refresh-scope-cmd';
 import CatScope from './commands/private-cmds/cat-scope-cmd';
 import ScopeConfig from './commands/public-cmds/scope-config-cmd';
+import Bind from './commands/public-cmds/bind-cmd';
 
 export default function registerCommands(): CommandRegistrar {
   return new CommandRegistrar(BIT_USAGE, BIT_DESCRIPTION, BIT_VERSION, [
@@ -58,5 +59,6 @@ export default function registerCommands(): CommandRegistrar {
     new RefreshScope(),
     new CatScope(),
     new ScopeConfig(),
+    new Bind(),
   ]);
 }