+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- add bind command (which calls the driver bind command)
+
 ## [0.5.10] - 2017-05-11 rc
 
 - fix bug with specs that need compiling for server use